checking code for flyout error
