+++ blocks/link-list/link-list.js
@@ -1,4 +1,4 @@
-export function toggleLinkListDetail(e) {
+function toggleLinkListDetail(e) {
   const { target } = e;
   if (window.innerWidth < 768) {
     if (target.classList.contains('expand')) {
@@ -27,6 +27,25 @@ export function generateLinkListDom(block) {
   linkListRTEElem.innerHTML = linkListDetail.innerHTML;
 
   return [linkListTitleElem, linkListRTEElem];
+}
+
+export function generateHeaderLinkList(props) {
+  const [a, b, ulList] = props;
+  console.log(a);
+  console.log(b);
+  console.log(ulList);
+  const menuteaserDOM = document.createRange().createContextualFragment(`
+      <div class="flyout-link-list">     
+        <div class ="link-list-wrapper vertical">
+          <div class="link-list block" data-block-name="link-list" data-block-status="loaded">
+            <h3 class="link-list-title">${b.textContent}</h3>
+              <div class="link-list-detail">
+               ${ulList.innerHTML}
+              </div>
+        </div>
+      </div>
+    `);
+  return menuteaserDOM;
 }
 
 export default function decorate(block) {
